refact: recommendApplicationService
